API

MetidaBioeq.bioequivalence

MetidaBioeq.bioequivalence — Function

bioequivalence(data;
    vars = nothing,
    subject::Union{String, Symbol},
    period::Union{String, Symbol, Nothing} = nothing,
    formulation::Union{String, Symbol},
    sequence::Union{String, Symbol, Nothing} = nothing,
    stage::Union{String, Symbol, Nothing} = nothing,
    reference::Union{String, Symbol, Nothing} = nothing,
    design::Union{String, Symbol, Nothing} = nothing,
    io::IO = stdout,
    seqcheck::Bool = true,
    designcheck::Bool = true,
    dropcheck::Bool = true,
    dropmissingsubj = false,
    dropincompletesubj = false,
    info::Bool = true,
    warns::Bool = true,
    autoseq::Bool = false,
    logt::Bool = true)

vars - variabel's column(s);

subject - subject's column;

period - period's column;

formulation - formulation's column;

sequence -sequence's column;

stage - stage's column;

reference - reference value for formulation column;

design - design: "parallel", "2X2", "2X2X2", "2X2X4", ets. (formulations X sequences X periods);

seqcheck - check sequencs;

designcheck - check design correctness;

dropcheck - dropuot check;

dropmissingsubj - drop subjects with missing data;

dropincompletesubj - drop subjects with no full sequence data (work only if seqcheck = true);

info - show information;

warns - show warnings;

autoseq - try to make sequence collumn;

logt - if true (default) data is already log-transformed, else log() will be used.

If dropmissingsubj or dropincompletesubj used - copy of the data will be filtered.

MetidaBioeq.estimate

MetidaBioeq.estimate — Function

estimate(be; estimator = "auto", method = "auto", supresswarn = false)

method - Model settings.

if method == "auto" than method A used for "2X2" and "2X2X2" designes, method P for "parallel" design and method B for any other.

Methods:

A using GLM and model @formula(var ~ formulation + period + sequence + subject)

B using MixedModels and model @formula(var ~ formulation + period + sequence + (1|subject)) or Metida and model @lmmformula(v ~ formulation + period + sequence, random = 1|subject:SI)

C using Metida and model @lmmformula(v ~ formulation + period + sequence, random = formulation|subject:CSH, repeated = formulation|subject:DIAG)

P using GLM and model @formula(var ~ formulation)

estimator - Estimator settings.

if estimator == "auto" than GLM used for "parallel" design; for "2X2" design used GLM if no droputs and MixedModels if dropout == true; for other designes with method C Metida used and MixedModel for other cases.

Estimators:

"glm" for GLM (https://juliastats.org/GLM.jl/stable/)

"mm" for MixedModels (https://juliastats.org/MixedModels.jl/stable/)

"met" for Metida (https://pharmcat.github.io/Metida.jl/stable/)

Other autosettings:

If design is "parallel" estimator set as "glm" and method as "P".

If design is "2X2" and method is "P" or "C" than if estimator == "glm" method set as "A" and "B" for other estimators.

If design not "parallel" or "2X2":

if method not "A", "B" or "C" than set as "A" for "glm" ann as B for other estimators;

if estimator == "glm" and method == "B" than estimator set as "mm", if estimator == "glm" or "mm" and method == "C" than estimator set as "met".

Reference:

EMA: GUIDELINE ON THE INVESTIGATION OF BIOEQUIVALENCE

EMA: GUIDELINE ON THE INVESTIGATION OF BIOEQUIVALENCE, Annex I

MetidaBioeq.result

MetidaBioeq.result — Function

result(beres::BEResults)

Returns dataframe with bioequivalence results.

MetidaBioeq.makeseq

MetidaBioeq.makeseq — Function

makeseq(data;
    subject = :subject,
    period = :period,
    formulation = :formulation)

Make sequence vector from data and subject, period, formulation columns.

MetidaBase

See https://github.com/PharmCat/MetidaBase.jl