API

MetidaBioeq.bioequivalence

MetidaBioeq.bioequivalence — Function

bioquivalence(data;
vars = nothing,
subject = :subject,
period = :period,
formulation = :formulation,
sequence = :sequence,
stage = nothing,
reference = nothing,
design = nothing,
io::IO = stdout,
seqcheck = true,
dropcheck = true,
logt = true)

vars - variabel's column(s);

subject - subject's column;

period - period's column;

formulation - formulation's column;

sequence -sequence's column;

stage - stage's column;

reference - reference value for formulation column;

design - design: "parallel", "2X2", "2X2X2", "2X2X4", ets. (formulations X sequences X periods);

seqcheck - check sequencs;

dropcheck - dropuot check;

logt - if true (default) data is already log-transformed, else log() will be used.

MetidaBioeq.makeseq

MetidaBioeq.makeseq — Function

makeseq(data;
    subject = :subject,
    period = :period,
    formulation = :formulation)

Make sequence vector from data and subject, period, formulation columns.

MetidaBioeq.result

MetidaBioeq.result — Function

result(be; estimator = "auto", method = "auto", supresswarn = false)

method - Model settings.

if method == "auto" than method A used for "2X2" and "2X2X2" designes, method P for "parallel" design and method B for any other.

Methods:

A using GLM and model @formula(var ~ formulation + period + sequence + subject)

B using MixedModels and model @formula(var ~ formulation + period + sequence + (1|subject)) or Metida and model @lmmformula(v ~ formulation + period + sequence, random = 1|subject:SI)

C using Metida and model @lmmformula(v ~ formulation + period + sequence, random = formulation|subject:CSH, repeated = formulation|subject:DIAG)

P using GLM and model @formula(var ~ formulation)

estimator - Estimator settings.

if estimator == "auto" than GLM used for "parallel" design; for "2X2" design used GLM if no droputs and MixedModels if dropout == true; for other designes with method C Metida used and MixedModel for other cases.

Estimators:

"glm" for GLM (https://juliastats.org/GLM.jl/stable/)

"mm" for MixedModels (https://juliastats.org/MixedModels.jl/stable/)

"met" for Metida (https://pharmcat.github.io/Metida.jl/stable/)

Other autosettings:

If design is "parallel" estimator set as "glm" and method as "P".

If design is "2X2" and method is "P" or "C" than if estimator == "glm" method set as "A" and "B" for other estimators.

If design not "parallel" or "2X2":

if method not "A", "B" or "C" than set as "A" for "glm" ann as B for other estimators;

if estimator == "glm" and method == "B" than estimator set as "mm", if estimator == "glm" or "mm" and method == "C" than estimator set as "met".

Reference:

EMA: GUIDELINE ON THE INVESTIGATION OF BIOEQUIVALENCE

EMA: GUIDELINE ON THE INVESTIGATION OF BIOEQUIVALENCE, Annex I